Update week-1-08262021.md
Thursday Notes, Aug 26, 2021
We went over the MITRE Attack Framework and MITRE Kill Chain
We then did a walkthrough of SimpleCTF in TryHackMe
We did an nmap scan of the network and found 3 or so ports that were open. One of which included SSH. 
We found a SQLi vulnerability that we worked on exploiting. CVE-2019-9053
We found a text file with a password. 
We ssh'd into the machine and found the next flag which was G00d j0b, keep up!
There was another user in the directory called sunbath
We leveraged vim to spawn a privileged shell and got the root flag. which was W3ll d0n3. You made it!
Official walk through was here. https://www.youtube.com/watch?v=mItfrdyE1U4



References
MITRE Attack Framework: https://attack.mitre.org/
MITRE Kill Chain: https://www.klogixsecurity.com/blog/strategic-guide-to-mitre-attck-framework